--- OpensslPkg/Library/BaseCryptLib/Hash/CryptParallelHashNull.c
@@ -1,40 +1,0 @@
-/** @file
-  ParallelHash Implementation which does not provide real capabilities.
-
-Copyright (c) 2022, Intel Corporation. All rights reserved.<BR>
-SPDX-License-Identifier: BSD-2-Clause-Patent
-
-**/
-
-#include "InternalCryptLib.h"
-
-/**
-  Parallel hash function ParallelHash256, as defined in NIST's Special Publication 800-185,
-  published December 2016.
-
-  @param[in]   Input            Pointer to the input message (X).
-  @param[in]   InputByteLen     The number(>0) of input bytes provided for the input data.
-  @param[in]   BlockSize        The size of each block (B).
-  @param[out]  Output           Pointer to the output buffer.
-  @param[in]   OutputByteLen    The desired number of output bytes (L).
-  @param[in]   Customization    Pointer to the customization string (S).
-  @param[in]   CustomByteLen    The length of the customization string in bytes.
-
-  @retval FALSE  This interface is not supported.
-
-**/
-BOOLEAN
-EFIAPI
-ParallelHash256HashAll (
-  IN CONST VOID   *Input,
-  IN       UINTN  InputByteLen,
-  IN       UINTN  BlockSize,
-  OUT      VOID   *Output,
-  IN       UINTN  OutputByteLen,
-  IN CONST VOID   *Customization,
-  IN       UINTN  CustomByteLen
-  )
-{
-  ASSERT (FALSE);
-  return FALSE;
-}
