--- OpensslPkg/Library/BaseCryptLib/Hash/CryptSha1.c
@@ -1,228 +1,0 @@
-/** @file
-  SHA-1 Digest Wrapper Implementation over OpenSSL.
-
-Copyright (c) 2009 - 2016, Intel Corporation. All rights reserved.<BR>
-SPDX-License-Identifier: BSD-2-Clause-Patent
-
-**/
-
-#include "InternalCryptLib.h"
-#include <openssl/sha.h>
-
-#ifndef DISABLE_SHA1_DEPRECATED_INTERFACES
-
-/**
-  Retrieves the size, in bytes, of the context buffer required for SHA-1 hash operations.
-
-  @return  The size, in bytes, of the context buffer required for SHA-1 hash operations.
-
-**/
-UINTN
-EFIAPI
-Sha1GetContextSize (
-  VOID
-  )
-{
-  //
-  // Retrieves OpenSSL SHA Context Size
-  //
-  return (UINTN)(sizeof (SHA_CTX));
-}
-
-/**
-  Initializes user-supplied memory pointed by Sha1Context as SHA-1 hash context for
-  subsequent use.
-
-  If Sha1Context is NULL, then return FALSE.
-
-  @param[out]  Sha1Context  Pointer to SHA-1 context being initialized.
-
-  @retval TRUE   SHA-1 context initialization succeeded.
-  @retval FALSE  SHA-1 context initialization failed.
-
-**/
-BOOLEAN
-EFIAPI
-Sha1Init (
-  OUT  VOID  *Sha1Context
-  )
-{
-  //
-  // Check input parameters.
-  //
-  if (Sha1Context == NULL) {
-    return FALSE;
-  }
-
-  //
-  // OpenSSL SHA-1 Context Initialization
-  //
-  return (BOOLEAN)(SHA1_Init ((SHA_CTX *)Sha1Context));
-}
-
-/**
-  Makes a copy of an existing SHA-1 context.
-
-  If Sha1Context is NULL, then return FALSE.
-  If NewSha1Context is NULL, then return FALSE.
-
-  @param[in]  Sha1Context     Pointer to SHA-1 context being copied.
-  @param[out] NewSha1Context  Pointer to new SHA-1 context.
-
-  @retval TRUE   SHA-1 context copy succeeded.
-  @retval FALSE  SHA-1 context copy failed.
-
-**/
-BOOLEAN
-EFIAPI
-Sha1Duplicate (
-  IN   CONST VOID  *Sha1Context,
-  OUT  VOID        *NewSha1Context
-  )
-{
-  //
-  // Check input parameters.
-  //
-  if ((Sha1Context == NULL) || (NewSha1Context == NULL)) {
-    return FALSE;
-  }
-
-  CopyMem (NewSha1Context, Sha1Context, sizeof (SHA_CTX));
-
-  return TRUE;
-}
-
-/**
-  Digests the input data and updates SHA-1 context.
-
-  This function performs SHA-1 digest on a data buffer of the specified size.
-  It can be called multiple times to compute the digest of long or discontinuous data streams.
-  SHA-1 context should be already correctly initialized by Sha1Init(), and should not be finalized
-  by Sha1Final(). Behavior with invalid context is undefined.
-
-  If Sha1Context is NULL, then return FALSE.
-
-  @param[in, out]  Sha1Context  Pointer to the SHA-1 context.
-  @param[in]       Data         Pointer to the buffer containing the data to be hashed.
-  @param[in]       DataSize     Size of Data buffer in bytes.
-
-  @retval TRUE   SHA-1 data digest succeeded.
-  @retval FALSE  SHA-1 data digest failed.
-
-**/
-BOOLEAN
-EFIAPI
-Sha1Update (
-  IN OUT  VOID        *Sha1Context,
-  IN      CONST VOID  *Data,
-  IN      UINTN       DataSize
-  )
-{
-  //
-  // Check input parameters.
-  //
-  if (Sha1Context == NULL) {
-    return FALSE;
-  }
-
-  //
-  // Check invalid parameters, in case that only DataLength was checked in OpenSSL
-  //
-  if ((Data == NULL) && (DataSize != 0)) {
-    return FALSE;
-  }
-
-  //
-  // OpenSSL SHA-1 Hash Update
-  //
-  return (BOOLEAN)(SHA1_Update ((SHA_CTX *)Sha1Context, Data, DataSize));
-}
-
-/**
-  Completes computation of the SHA-1 digest value.
-
-  This function completes SHA-1 hash computation and retrieves the digest value into
-  the specified memory. After this function has been called, the SHA-1 context cannot
-  be used again.
-  SHA-1 context should be already correctly initialized by Sha1Init(), and should not be
-  finalized by Sha1Final(). Behavior with invalid SHA-1 context is undefined.
-
-  If Sha1Context is NULL, then return FALSE.
-  If HashValue is NULL, then return FALSE.
-
-  @param[in, out]  Sha1Context  Pointer to the SHA-1 context.
-  @param[out]      HashValue    Pointer to a buffer that receives the SHA-1 digest
-                                value (20 bytes).
-
-  @retval TRUE   SHA-1 digest computation succeeded.
-  @retval FALSE  SHA-1 digest computation failed.
-
-**/
-BOOLEAN
-EFIAPI
-Sha1Final (
-  IN OUT  VOID   *Sha1Context,
-  OUT     UINT8  *HashValue
-  )
-{
-  //
-  // Check input parameters.
-  //
-  if ((Sha1Context == NULL) || (HashValue == NULL)) {
-    return FALSE;
-  }
-
-  //
-  // OpenSSL SHA-1 Hash Finalization
-  //
-  return (BOOLEAN)(SHA1_Final (HashValue, (SHA_CTX *)Sha1Context));
-}
-
-/**
-  Computes the SHA-1 message digest of a input data buffer.
-
-  This function performs the SHA-1 message digest of a given data buffer, and places
-  the digest value into the specified memory.
-
-  If this interface is not supported, then return FALSE.
-
-  @param[in]   Data        Pointer to the buffer containing the data to be hashed.
-  @param[in]   DataSize    Size of Data buffer in bytes.
-  @param[out]  HashValue   Pointer to a buffer that receives the SHA-1 digest
-                           value (20 bytes).
-
-  @retval TRUE   SHA-1 digest computation succeeded.
-  @retval FALSE  SHA-1 digest computation failed.
-  @retval FALSE  This interface is not supported.
-
-**/
-BOOLEAN
-EFIAPI
-Sha1HashAll (
-  IN   CONST VOID  *Data,
-  IN   UINTN       DataSize,
-  OUT  UINT8       *HashValue
-  )
-{
-  //
-  // Check input parameters.
-  //
-  if (HashValue == NULL) {
-    return FALSE;
-  }
-
-  if ((Data == NULL) && (DataSize != 0)) {
-    return FALSE;
-  }
-
-  //
-  // OpenSSL SHA-1 Hash Computation.
-  //
-  if (SHA1 (Data, DataSize, HashValue) == NULL) {
-    return FALSE;
-  } else {
-    return TRUE;
-  }
-}
-
-#endif
